+++ index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8" />
     <link rel="icon" type="png" href="/src/assets/Minilogo.png" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>MsgCraft</title>
+    <title>MessBee</title>
   </head>
   <body>
     <div id="root"></div>
