+++ package.json
@@ -1,5 +1,6 @@
 {
-  "name": "msgcraft",
+  "name": "messbee",
+  "home":"./",
   "private": true,
   "version": "0.0.0",
   "type": "module",
