+++ src/components/chat-inbox/ChatInbox.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const ChatInbox = () => {
+  return (
+    <div><p>hello world</p></div>
+  )
+}
+
+export default ChatInbox
