+++ src/components/header/FirstHeader.css
@@ -1,0 +1,18 @@
+.first-heading {
+  background-color: #F0B429;
+  width: 100%;
+  height: 1.5rem;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.sub-heading {
+  margin: 0 10px;
+  font-size: 0.7rem;
+  font-weight: 500;
+  color: #610404;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+  overflow: hidden;
+}/*# sourceMappingURL=FirstHeader.css.map */
