+++ src/components/mainsidebar/MainSidebar.css
@@ -1,11 +1,12 @@
 .mainSideBar {
   width: 5rem;
-  min-height: 100vh;
+  height: 100vh;
   position: relative;
-  overflow: hidden;
-  padding: 1.5rem 1.5rem 1.5rem 1.5rem;
+  overflow: auto;
+  padding: 0.5rem;
   background-color: #FAF9F7;
   display: flex;
+  flex-wrap: nowrap;
   flex-direction: column;
   align-items: center;
   justify-content: space-between;
@@ -20,7 +21,7 @@
 }
 
 .logo {
-  width: 3.5rem;
+  width: 3.7rem;
   cursor: pointer;
   overflow: hidden;
 }
@@ -57,8 +58,14 @@
 }
 
 .menuContener1 {
-  margin-top: 1rem;
+  margin-top: 2rem;
   display: flex;
   flex-direction: column;
   gap: 0.2rem;
+}
+
+@media (max-width: 600) {
+  .mainSideBar {
+    overflow: hidden;
+  }
 }/*# sourceMappingURL=MainSidebar.css.map */
