+++ src/components/mainsidebar/MainSidebar.scss
@@ -1,11 +1,12 @@
 .mainSideBar{
-  width: 5rem;
-  min-height: 100vh;
+  min-width: 5rem;
+  height: 100vh;
   position: relative;
-  overflow: hidden;
-  padding: 1.5rem 1.5rem 1.5rem 1.5rem;
+  overflow-y: scroll;
+  padding: 0.5rem ;
   background-color: #FAF9F7; 
   display: flex;
+  flex-wrap: nowrap;
   flex-direction: column;
   align-items: center;
   justify-content: space-between;
@@ -20,7 +21,7 @@
 
 }
 .logo{
-  width: 3.5rem;
+  width: 3.7rem;
   cursor: pointer;
   overflow: hidden;
 }
@@ -55,8 +56,14 @@
   gap: 0.2rem;
 }
 .menuContener1{
-  margin-top: 1rem;
+  margin-top: 2rem;
   display: flex;
   flex-direction: column;
   gap: 0.2rem;
+}
+
+@media (max-width:600) {
+  .mainSideBar{
+   overflow: hidden;  
+  }
 }
