+++ src/pages/campaign/campaign.css
@@ -1,0 +1,10 @@
+.main {
+  display: flex;
+  width: 100%;
+}
+
+.body {
+  width: 100%;
+  height: 1100px;
+  background-color: green;
+}/*# sourceMappingURL=campaign.css.map */
