+++ src/pages/campaign/campaign.jsx
@@ -1,12 +1,13 @@
 import React from "react";
-import MainSidebar from "../components/mainsidebar/MainSidebar";
-import FirstHeader from "../components/header/FirstHeader";
-import MainHeading from "../components/header/MainHeading";
+import MainSidebar from "../../components/mainsidebar/MainSidebar";
+import FirstHeader from "../../components/header/FirstHeader";
+import MainHeading from "../../components/header/MainHeading";
+import "./campaign.scss"
 
 
 const campaign = () => {
   return (
-    <div className="maindashboard">
+    <div className="main">
       <MainSidebar />
       <div className="body">
         <FirstHeader></FirstHeader>
