+++ src/pages/campaign/campaign.scss
@@ -1,0 +1,10 @@
+.main{
+  display: flex;
+  width: 100%;
+  // min-height: 100vh;
+}
+.body{
+  width: 100%;
+  height: 1100px;
+  background-color: green;
+}
