+++ src/pages/chat/chat.scss
@@ -1,13 +1,10 @@
-@import './mixin';
-.maindashboard{
+.main{
   display: flex;
   width: 100%;
-  
   // min-height: 100vh;
-  
-
 }
 .body{
-  width: 100vw;
- 
+  width: 100%;
+  height: 1100px;
+  background-color: red;
 }
