+++ src/pages/contats/contact.css
@@ -1,0 +1,10 @@
+.main {
+  display: flex;
+  width: 100%;
+}
+
+.body {
+  width: 100%;
+  height: 1100px;
+  background-color: blue;
+}/*# sourceMappingURL=contact.css.map */
