+++ src/pages/help/help.jsx
@@ -1,0 +1,20 @@
+import React from "react";
+import MainSidebar from "../../components/mainsidebar/MainSidebar";
+import FirstHeader from "../../components/header/FirstHeader";
+import MainHeading from "../../components/header/MainHeading";
+import "./help.scss"
+
+
+const help = () => {
+  return (
+    <div className="main">
+      <MainSidebar />
+      <div className="body">
+        <FirstHeader></FirstHeader>
+        <MainHeading></MainHeading>
+      </div>
+    </div>
+  );
+};
+
+export default help;
