+++ src/styles/dashboard-paid.css
@@ -1,0 +1,9 @@
+@import url("https://fonts.googleapis.com/css2?family=Poppins:ital,wght@0,300;0,400;0,500;0,600;0,800;1,500;1,600&display=swap");
+.maindashboard {
+  display: flex;
+  width: 100%;
+}
+
+.body {
+  width: 100vw;
+}/*# sourceMappingURL=dashboard-paid.css.map */
